requirements.txt
langchain
langchain-groq
langchain-community
langchain-openai
pinecone-client
streamlit
python-dotenv
pypdf2
tiktoken
langdetect